Move stream related code to Streams directory
