+++ Archives/ClmFile.h
@@ -2,8 +2,8 @@
 
 #include "WaveFile.h"
 #include "ArchiveFile.h"
-#include "../StreamReader.h"
-#include "../StreamWriter.h"
+#include "../Streams/StreamReader.h"
+#include "../Streams/StreamWriter.h"
 #include <cstdint>
 #include <string>
 #include <vector>
